+++ app/src/main/java/ocrreader/injection/Application.kt
@@ -1,6 +1,7 @@
-package ocrreader
+package ocrreader.injection
 
 import android.app.Application
+import ocrreader.DaggerContextComponent
 
 class EdGridApplication : Application() {
     lateinit var component: ContextComponent
